fix frame number

+++ src/render/frame_data.rs
@@ -4,7 +4,7 @@ pub struct FrameData {
     pub w: usize,
     pub h: usize,
     pub pixels: Vec<Color>,
-    pub frame_number: i32,
+    pub frame_number: usize,
     pub t: f64,
     // TODO: camera settings (for debugging)
 }
